0x0F-python-object_relational_mapping

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -4,18 +4,18 @@
 import sys
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from model_state import Base, State
+from model_state import State
 
 if __name__ == "__main__":
     # Create a new database engine using provided credentials
-    engine = create_engine('mysql+mysqldb://{}:{}@localhost:3306/{}'
-                           .format(sys.argv[1], sys.argv[2], sys.argv[3]))
+    engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}"
+                           .format(sys.argv[1], sys.argv[2], sys.argv[3]),
+                           pool_pre_ping=True)
     Session = sessionmaker(bind=engine)
     session = Session()
-    
-    # Query all State objects with names containing the letter 'a'
-    states = session.query(State).filter(State.name.like('%a%')).all()
-    # Delete each state object retrieved from the query
-    for state in states:
-        session.delete(state)
+
+    # Query and delete State objects with names containing the letter 'a'
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)  # Delete the state object
     session.commit()  # Commit the changes to the database
